iter9 рефакторинг по комментариям ревью
